Eliminate various issues brought up by scan-build.
Acked-by: Stefan Reinauer <stepan@coresystems.de>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@4152 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/devices/pci_rom.c
@@ -42,7 +42,7 @@ struct rom_header * pci_rom_probe(struct device *dev)
 		printk_debug("In cbfs, rom address for %s = %lx\n", 
 				dev_path(dev), rom_address);
 		if (v) {
-			dev->rom_address = v;
+			dev->rom_address = (u32)v;
 			dev->on_mainboard = 1;
 		}
 	}
